+++ thonnycontrib/gitonic/__init__.py
@@ -1,24 +1,49 @@
+from thonny.ui_utils import select_sequence
+from thonny import get_workbench
+from tkinter import messagebox
 import os
+import configparser
 
 VERSION = "v0.0.3-a"
 
 # import subprocess
 
-from tkinter import messagebox
 
-from thonny import get_workbench
-from thonny.ui_utils import select_sequence
+def open_config():
+    cfg = configparser.ConfigParser(allow_no_value=True)
+    cfg.read_dict({"DEFAULT": {"path": "."}})
+
+    fnam = os.path.join("~", ".gitonic", "thonnycontrib.cfg")
+    fnam = os.path.expanduser(fnam)
+
+    try:
+        with open(fnam) as f:
+            cfg.read_string(f.read())
+    except Exception as ex:
+        print("defaulting. file not found.", ex)
+        messagebox.showerror("internal error", fnam + " " + str(ex))
+    return cfg
 
 
 def open_gitonic():
+
+    cfg = open_config()
+    fp = cfg["DEFAULT"]["path"]
+
+    fp = os.path.join(fp, "gitonic")
+    fp = os.path.expanduser(fp)
+
+    fnam = fp + " &"
+
     try:
         # this will log to the same console like thonny
         # todo
-        os.system("gitonic &")
+        print("running",fnam)
+        os.system(fnam)
         # subprocess.Popen(["gitonic"],creationflags=0)
     except Exception as ex:
-        print(ex)
-        messagebox.showerror("internal error", "could not start gitonic")
+        messagebox.showerror(
+            "internal error", "could not start gitonic " + str(ex))
 
 
 def load_plugin():
@@ -32,4 +57,5 @@ def load_plugin():
         )
     except Exception as ex:
         print(ex)
-        messagebox.showerror("internal error", "could not start gitonic plugin")
+        messagebox.showerror(
+            "internal error", "could not start gitonic plugin")